refactor(api): extract owner write guards to write_validation.py
